# Remove invalid optimization from put that negatively impacts performance