@trapexit 2017-06-06
Do you mean "decreases flag memory usage"?