@dwforbes 2017-06-06
The first commit -- 9b12aa4 -- was the only one I intended to get captured in the pull request, in that case removing an unnecessary extra conditional that adds overhead. In that case I was capturing some quick analysis I had done out of curiosity regarding the performance of C in the nucleotide benchmark game. It is a change that can only be positive, even for projects that might have dependencies on internals.

The second commit -- 348601f -- is a potentially breaking change that reduces the amount of bit packing, increasing the memory footprint for flags (albeit still a tiny overhead for all but the most enormous of hash tables), however it has a dramatic benefit to performance on all architectures because it removes three expensive bit shifts be evaluation. It can break users who might have hard-fixed dependencies on the structure of flags, for instance, so it was more an example of cost-benefit analysis of memory versus performance.